+++ integreat_cms/static/src/js/media-management/library.tsx
@@ -2,13 +2,13 @@
  * This component renders the media library in edit mode,
  * so new directories and files can be added and the existing entries can be modified
  */
-import { FilePlus, FolderPlus, Search, Loader } from "preact-feather";
+import { FilePlus, FolderPlus, Search, Loader } from "lucide-preact";
 import { StateUpdater, useEffect, useState } from "preact/hooks";
 import { route } from "preact-router";
 
 import { Directory, MediaApiPaths, MediaLibraryEntry, File } from ".";
 import Breadcrumbs from "./component/breadcrumbs";
-import DirectoryContent from "./component/directory-content";
+import DirectoryContent, { DraggedElement } from "./component/directory-content";
 import EditDirectorySidebar from "./component/edit-directory-sidebar";
 import EditSidebar from "./component/edit-sidebar";
 import { Message } from "./component/message";
@@ -82,6 +82,15 @@ export default function Library({
   // Whether or not the file upload form should be shown
   const [isUploadFile, setUploadFile] = useState<boolean>(false);
 
+  const [draggedItem, setDraggedItem] = useState<DraggedElement | null>(null);
+
+  const moveIntoDirectory = async (targetDirectoryId: number) => {
+    let form = document.getElementById("media-move-form") as HTMLFormElement;
+    (document.getElementById("parent_directory") as HTMLInputElement).value =
+      targetDirectoryId == 0 ? "" : targetDirectoryId.toString();
+    (document.getElementById("media-move-btn") as HTMLInputElement).click();
+  };
+
   // This submit function is used for all form submissions
   const submitForm = async (event: Event, successCallback?: (data: any) => void) => {
     event.preventDefault();
@@ -97,9 +106,9 @@ export default function Library({
         },
         body: new FormData(form),
       });
-      console.log(response);
+      console.debug(response);
       const data = await response.json();
-      console.log(data);
+      console.debug(data);
       if (response.status === 200) {
         console.debug("Form submission successful!");
         if (typeof successCallback === "function") {
@@ -161,6 +170,18 @@ export default function Library({
       <h1 className="w-full heading p-2">{mediaTranslations.heading_media_library}</h1>
       <div className="flex flex-wrap justify-between gap-x-2 gap-y-4">
         <div id="table-search" class="flex">
+          <form
+            id="media-move-form"
+            class="hidden"
+            method="POST"
+            encType="multipart/form-data"
+            onSubmit={submitForm}
+            action={apiEndpoints.moveFile}
+          >
+            <input name="mediafile_id" value={draggedItem?.id} />
+            <input name="parent_directory" id="parent_directory" />
+            <input id="media-move-btn" type="submit" />
+          </form>
           <form
             id="media-search-form"
             class="relative"
@@ -260,6 +281,8 @@ export default function Library({
                 breadCrumbs={directoryPath}
                 searchQuery={searchQuery}
                 mediaTranslations={mediaTranslations}
+                allowDrop={draggedItem !== null}
+                dropItem={moveIntoDirectory}
               />
             </div>
             {isLoading ? (
@@ -271,6 +294,9 @@ export default function Library({
                   mediaLibraryContent={mediaLibraryContent}
                   mediaTranslations={mediaTranslations}
                   globalEdit={globalEdit}
+                  allowDrop={draggedItem !== null}
+                  setDraggedItem={setDraggedItem}
+                  dropItem={moveIntoDirectory}
                 />
               </div>
             )}